requirements.txt
#api==1.0.0
asn1crypto==0.24.0
aspy.yaml==1.3.0
astroid==2.5.3
attrs==19.1.0
beautifulsoup4==4.6.3
cached-property==1.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
cryptography==2.3.1
cycler==0.10.0
dask==2.4.0
decorator==4.4.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.4
django-bower==5.2.0
django-cors-headers==3.2.1
django-datetime-widget==0.9.3
django-debug-toolbar==2.1
django-extensions==2.2.9
django-filter==2.2.0
django-happenings==0.3.3
django-mysql==3.3.0
django-npm==1.0.0
django-scheduler==0.9.3
djangorestframework==3.11.0
entrypoints==0.3
falcon==1.4.1
flake8==3.7.9
Flask==1.1.2
gunicorn==19.9.0
httpie==2.1.0
icalendar==4.0.4
idna==2.7
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.0.2
kiwisolver==1.1.0
lazy-object-proxy==1.6.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mysqlclient==1.4.6
networkx==2.3
nodeenv==1.3.3
nose==1.3.7
npm==0.1.1
numpy==1.16.0
opencv-contrib-python==4.0.0.21
optional-django==0.1.0
pika==1.1.0
Pillow==6.1.0
pre-commit==0.8.0
preppy==3.0.1
psycopg2==2.8.6
py2exe==0.9.2.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.7.4
pyparsing==2.4.2
PyPDF2==1.26.0
pyrsistent==0.15.4
pyterminalsize==0.1.0
python-dateutil==2.8.0
python-decouple==3.4
python-mimeparse==1.6.0
python-nmap==0.6.1
pytz==2018.5
PyWavelets==1.0.3
PyYAML==5.1.2
reportlab==3.5.66
requests==2.23.0
scikit-image==0.14.2
scikit-learn==0.24.1
scipy==1.3.1
six==1.11.0
sqlparse==0.3.0
static3==0.7.0
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.10.0
typed-ast==1.4.3
urllib3==1.23
virtualenv==16.7.4
Werkzeug==1.0.1
whitenoise==4.1.2
wrapt==1.12.1
